# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{DEV} Pedia",
        "title": "{DEV} Pedia API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:9001",
    "basePath": "/",
    "paths": {
...
        "/api/articles/{source_type}/{id}/backlinks": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all articles that link to the specified article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Get article backlinks",
                "parameters": [
                    {
                        "enum": [
                            "doc",
                            "git"
                        ],
                        "type": "string",
                        "description": "Source type",
                        "name": "source_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GetBacklinksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/articles/{source_type}/{id}/broken-links": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all broken wiki links (non-existent or archived targets) in the specified article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Get broken links in an article",
                "parameters": [
                    {
                        "enum": [
                            "doc",
                            "git"
                        ],
                        "type": "string",
                        "description": "Source type",
                        "name": "source_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GetBrokenLinksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/graph": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get the complete knowledge graph filtered by user's classification level",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Get full knowledge graph",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Minimum classification level",
                        "name": "min_classification",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum classification level",
                        "name": "max_classification",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated source types (doc,git)",
                        "name": "source_types",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only show hub nodes",
                        "name": "only_hubs",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only show authority nodes",
                        "name": "only_authorities",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only show orphan nodes",
                        "name": "only_orphans",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Exclude orphan nodes",
                        "name": "exclude_orphans",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GetGraphResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/graph/article/{source_type}/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a subgraph centered on a specific article (N-hop neighborhood)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Get article neighborhood graph",
                "parameters": [
                    {
                        "enum": [
                            "doc",
                            "git"
                        ],
                        "type": "string",
                        "description": "Source type",
                        "name": "source_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 2,
                        "description": "Depth of neighborhood (1-5)",
                        "name": "depth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GetGraphResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/graph/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get overall statistics about the knowledge graph",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Get graph statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GraphStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fceek_dev-pedia_backend_internal_models.ArticleSourceType": {
            "type": "string",
            "enum": [
                "doc",
                "git"
            ],
            "x-enum-varnames": [
                "ArticleSourceDoc",
                "ArticleSourceGit"
            ]
        },
        "fceek_dev-pedia_backend_internal_models.ArticleStatus": {
            "type": "string",
            "enum": [
                "draft",
                "published",
                "archived"
            ],
            "x-enum-varnames": [
                "ArticleStatusDraft",
                "ArticleStatusPublished",
                "ArticleStatusArchived"
            ]
        },
        "fceek_dev-pedia_backend_internal_models.ArticleTag": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.BacklinkSummary": {
            "type": "object",
            "properties": {
                "context_snippet": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "link_text": {
                    "type": "string"
                },
                "source_article_id": {
                    "type": "string"
                },
                "source_article_type": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleSourceType"
                },
                "source_classification": {
                    "type": "integer"
                },
                "source_path": {
                    "type": "string"
                },
                "source_title": {
                    "type": "string"
                }
            }
        },
        "fceek_dev-pedia_backend_internal_models.BrokenLink": {
            "type": "object",
            "properties": {
                "end_position": {
                    "description": "Character position in content",
                    "type": "integer"
                },
                "link_text": {
                    "description": "The full [[...]] text",
                    "type": "string"
                },
                "reason": {
                    "description": "Why the link is broken",
                    "type": "string"
                },
                "start_position": {
                    "description": "Character position in content",
                    "type": "integer"
                },
                "target_path": {
                    "description": "The path that couldn't be resolved",
                    "type": "string"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.GetBacklinksResponse": {
            "type": "object",
            "properties": {
                "backlinks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.BacklinkSummary"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "fceek_dev-pedia_backend_internal_models.GetBrokenLinksResponse": {
            "type": "object",
            "properties": {
                "broken_links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.BrokenLink"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.GetGraphResponse": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GraphEdge"
                    }
                },
                "filtered_by": {
                    "description": "Description of applied filters",
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GraphNode"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GraphStats"
                },
                "user_classification": {
                    "description": "User's classification level for client-side filtering",
                    "type": "integer"
                }
            }
        },
        "fceek_dev-pedia_backend_internal_models.GraphEdge": {
            "type": "object",
            "properties": {
                "context_snippet": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "description": "link text",
                    "type": "string"
                },
                "source": {
                    "description": "source article ID",
                    "type": "string"
                },
                "target": {
                    "description": "target article ID",
                    "type": "string"
                },
                "type": {
                    "description": "'wiki', 'mention', 'embed'",
                    "type": "string"
                }
            }
        },
        "fceek_dev-pedia_backend_internal_models.GraphNode": {
            "type": "object",
            "properties": {
                "classification_level": {
                    "type": "integer"
                },
                "full_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inbound_count": {
                    "description": "Graph metrics",
                    "type": "integer"
                },
                "is_authority": {
                    "type": "boolean"
                },
                "is_hub": {
                    "type": "boolean"
                },
                "is_orphan": {
                    "type": "boolean"
                },
                "outbound_count": {
                    "type": "integer"
                },
                "source_type": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleSourceType"
                },
                "status": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleStatus"
                },
                "tags": {
                    "description": "Optional metadata",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleTag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "total_degree": {
                    "type": "integer"
                }
            }
        },
        "fceek_dev-pedia_backend_internal_models.GraphStats": {
            "type": "object",
            "properties": {
                "authorities_count": {
                    "type": "integer"
                },
                "average_degree": {
                    "type": "number"
                },
                "hubs_count": {
                    "type": "integer"
                },
                "max_degree": {
                    "type": "integer"
                },
                "nodes_by_classification": {
                    "description": "Classification breakdown",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "orphans_count": {
                    "type": "integer"
                },
                "total_edges": {
                    "type": "integer"
                },
                "total_nodes": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer token for regular authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GodToken": {
            "description": "God token for bootstrap operations",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}